+++ packages/services/f-statistics/src/defaultOptions.js
@@ -2,9 +2,9 @@ const defaultConfig = {
     namespace: 'f-statistics',
     environment: 'test',
     featureName: 'Generic Front End',
-    logsMaxByteSize: 1000,
+    logsMaxByteSize: 2000,
     logsMaxLength: 5,
-    logsIntervalTimer: 0
+    logsIntervalTimer: 5000
 };
 
 export default defaultConfig;